SCHED-99: Add new module structure and missing typing hints
